再利用性を高めるため、共通の関数をモジュールにする。
